pyproject.toml
[project]
name = "agentic-idea-factory"
version = "0.1.0"
description = "Python-orchestrated idea pipeline with MCP persona agents"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.5.0",
    "aiosqlite>=0.19.0",
    "httpx>=0.26.0",
    "anthropic>=0.18.0",
    "google-generativeai>=0.4.0",
    "google-auth>=2.0.0",
    "google-api-python-client>=2.0.0",
    "python-dotenv>=1.0.0",
    "PyJWT[crypto]>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
